Update Automated Progress Report Generator.py
Replaced pyautogui search for iDESK icon on desktop with python webbrowser call for url. This removes dependency on the location of the icon which at times appears off screen. This also allows me to change my screensaver without having to continuously update the iDESK Icon image which is impacted by the color changes in pixels of the background since it is semi-transparent.